Implement buddy allocation initialization.

+++ Kernel/include/common/string.h
@@ -1,34 +1,34 @@
 extern "C" {
-    void* memset(void* src, int c, size_t count);
-    void* memcpy(void* dest, const void* src, size_t count);
-    int memcmp(const void* s1, const void* s2, size_t n);
+void* memset(void* src, int c, unsigned long count);
+void* memcpy(void* dest, const void* src, unsigned long count);
+int memcmp(const void* s1, const void* s2, unsigned long n);
 
-    void strcpy(char* dest, const char* src);
-    void strncpy(char* dest, const char* src, size_t n);
-    int strcmp(const char* s1, const char* s2);
+void strcpy(char* dest, const char* src);
+void strncpy(char* dest, const char* src, unsigned long n);
+int strcmp(const char* s1, const char* s2);
 
-    char* strtok_r(char* str, const char* delim, char** saveptr);
+char* strtok_r(char* str, const char* delim, char** saveptr);
 
-    size_t strlen(const char* str);
-    char* strcat(char* dest, const char* src);
-    char* strncat(char* dest, const char* src, size_t n);
+unsigned long strlen(const char* str);
+char* strcat(char* dest, const char* src);
+char* strncat(char* dest, const char* src, unsigned long n);
 
-    int strncmp(const char* s1, const char* s2, size_t n);
+int strncmp(const char* s1, const char* s2, unsigned long n);
 
-    char* strupr(char* s);
-    char* strnupr(char* s, size_t n);
+char* strupr(char* s);
+char* strnupr(char* s, unsigned long n);
 
-    char* strchr(const char* s, int c);
-    char* strnchr(const char* s, int c, size_t n);
-    char* strrchr(const char* s, int c);
+char* strchr(const char* s, int c);
+char* strnchr(const char* s, int c, unsigned long n);
+char* strrchr(const char* s, int c);
 
-    char* strdup(const char* s);
+char* strdup(const char* s);
 }
 
 class String {
 private:
     int length;
-    char *str;
+    char* str;
 
 public:
     String() {
@@ -41,7 +41,8 @@ public:
         strcpy(this->str, _str);
     }
     ~String() {
-        if (this->str != nullptr) {
+        if (this->str != nullptr)
+        {
             delete[] this->str;
             this->str = nullptr;
         }
